agents/sentiment-agent/pyproject.toml
[project]
name = "sentiment-agent"
version = "1.0.0"
description = "A2A Sentiment Analysis Agent for financial news and market sentiment analysis"
authors = [
    { name = "AlphaAgents Financial", email = "info@alphaagents.finance" }
]
dependencies = [
    "langchain>=0.3.27,<0.4.0",
    "langchain-openai>=0.3.32,<0.4.0",
    "langchain-community>=0.3.28,<0.4.0",
    "openai>=1.102.0",
    "yfinance>=0.2.65",
    "pandas>=2.3.2",
    "numpy>=2.3.2",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "httpx>=0.28.1",
    "aiohttp>=3.12.15",
    "requests>=2.32.5",
    "beautifulsoup4>=4.13.5",
    "feedparser>=6.0.11",
    "newspaper3k>=0.2.8",
    "textblob>=0.18.0",
    "vaderSentiment>=3.3.2",
    "python-dateutil>=2.9.0",
    "langsmith>=0.2.0",
    "boto3>=1.35.0",
    "langchain-aws>=0.2.0,<0.3.0"
]
readme = "README.md"
requires-python = ">=3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "black>=25.1.0",
    "ruff>=0.12.10",
    "mypy>=1.17.1"
]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
select = ["E", "F", "I", "N", "W"]
ignore = []
line-length = 88
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true